backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pandas==2.2.3
openpyxl==3.1.5
google-generativeai==0.8.3
resend==2.3.0
python-dotenv==1.0.1
pydantic==2.9.2
